% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols.R
\name{get_symbols}
\alias{get_symbols}
\title{Get symbols}
\usage{
get_symbols(exchange = as.character(NA))
}
\arguments{
\item{exchange}{Name of an exchange (e.g. "binance", "kraken", "lykke")}
}
\value{
Character vector with symbols tradeable on specific exchange
}
\description{
This function scraps the symbols from the public API of a host of exchanges
}
